flask_crew_v4/flask-crewai-backend/requirements.txt
flask
requests
beautifulsoup4
PyGithub
crewai
crewai-tools
supabase
apscheduler
python-dotenv
